pyproject.toml
[tool.poetry]
name = "fastapizero"
version = "0.1.0"
description = ""
authors = ["yanndrade <ybamello@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = {extras = ["standard"], version = "^0.115.4"}
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.7.0"
alembic = "^1.14.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.19"
pyjwt = "^2.10.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
taskipy = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# <Ruff configuration>
[tool.ruff]
line-length = 79
extend-exclude = ["migrations"]

[tool.ruff.lint]
preview = true
select = ["I", "E", "W", "F", "PL", "PT"]

[tool.ruff.format]
preview = true
quote-style = "single"
# </Ruff configuration>

# <Pytest configuration>
[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
# </Pytest configuration>

# <Taskipy configuration>
[tool.taskipy.tasks]
run = "fastapi dev fastapizero/app.py"

#pre_test = 'task lint'
test = "pytest --cov=fastapizero -vv"
post_test = 'coverage html'

lint = "ruff check . && ruff check . --diff"
format = "ruff check . --fix && ruff format ."
# </Taskipy configuration>

